Widen the RollingCounter edge weighting to 64 bit
counts * inWindow is a 32 bit product, so a bucket holding more than
2^32 / BucketMs events wraps. At a 5 minute width that is about 14k: a
bucket of 50000 reported 11367 instead of 40000 once it reached the
window edge.

Below the threshold nothing changes, so lightning was unaffected, but
the helper is meant to be reused by counters with far higher rates.

test_large_burst_at_window_edge covers it. The existing burst test
sampled only inside the window, where the bucket is whole and never
weighted.

+++ src/modules/Telemetry/Sensor/RollingCounter.h
@@ -50,7 +50,9 @@ template <uint32_t WindowMs, uint32_t BucketMs> class RollingCounter
         // holds the total at exactly WindowMs as the current bucket fills.
         uint32_t elapsed = Time::getMillis() - bucketStartMs;
         uint32_t inWindow = elapsed < BucketMs ? BucketMs - elapsed : 0;
-        total += (counts[(head + 1) % BUCKETS] * inWindow + BucketMs / 2) / BucketMs;
+        // 64-bit: the product overflows 32 bits once a bucket holds more than 2^32 / BucketMs
+        // events, which is only ~14k at a 5 minute width.
+        total += (uint32_t)(((uint64_t)counts[(head + 1) % BUCKETS] * inWindow + BucketMs / 2) / BucketMs);
         return total;
     }
